requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.6
cryptography==41.0.2
deprecation==2.1.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg==3.1.14
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.6.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==0.21.1
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
testcontainers==3.7.1
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.1.0
uvicorn==0.22.0
wrapt==1.16.0
zipp==3.15.0
